Initial combine and simplification of functions
