+++ package.json
@@ -83,6 +83,7 @@
     "posttest": "eslint .",
     "prepack": "oclif-dev manifest && oclif-dev readme",
     "test": "nyc mocha --forbid-only -t 50000 \"test/**/*.test.js\"",
+    "test-create-wallet": "nyc mocha --forbid-only -t 50000 \"test/commands/create-wallet.test.js\"",
     "test-ots": "nyc mocha --forbid-only -t 50000 \"test/commands/ots.test.js\"",
     "test-receive": "nyc mocha --forbid-only -t 50000 \"test/commands/receive.test.js\"",
     "test-balance": "nyc mocha --forbid-only -t 50000 \"test/commands/balance.test.js\"",
